Powershell now primary
I'm now kicking off the powershell script, directly, rather than using the CMD file.

I'm still copying the files to the files to the Options folder so it can be kicked off manually, if desired, but I want the hard reboot command to kick off so the computer reboots to fully implement the uninstall.

Also added the HP Wolf Security to the  uninstall list.